+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [6.3.0](https://github.com/117/alpaca/compare/v6.2.7...v6.3.0) (2021-08-31)
+
 ### [6.2.7](https://github.com/117/alpaca/compare/v6.2.6...v6.2.7) (2021-08-31)
 
 ### [6.2.6](https://github.com/117/alpaca/compare/v6.2.5...v6.2.6) (2021-08-17)